Move BootTimes to l.lib.sysstats.
