+++ landscape/monitor/tests/test_computeruptime.py
@@ -1,113 +1,7 @@
-from datetime import datetime
-import struct
-
-from landscape.monitor.computeruptime import (
-        LoginInfo, LoginInfoReader, ComputerUptime, BootTimes)
+from landscape.lib.testing import append_login_data
+from landscape.monitor.computeruptime import ComputerUptime
 from landscape.tests.helpers import LandscapeTest, MonitorHelper
 from mock import ANY, Mock
-
-
-def append_login_data(filename, login_type=0, pid=0, tty_device="/dev/",
-                      id="", username="", hostname="", termination_status=0,
-                      exit_status=0, session_id=0, entry_time_seconds=0,
-                      entry_time_milliseconds=0,
-                      remote_ip_address=[0, 0, 0, 0]):
-    """Append binary login data to the specified filename."""
-    file = open(filename, "ab")
-    try:
-        file.write(struct.pack(LoginInfo.RAW_FORMAT, login_type, pid,
-                               tty_device.encode("utf-8"), id.encode("utf-8"),
-                               username.encode("utf-8"),
-                               hostname.encode("utf-8"),
-                               termination_status, exit_status, session_id,
-                               entry_time_seconds, entry_time_milliseconds,
-                               remote_ip_address[0], remote_ip_address[1],
-                               remote_ip_address[2], remote_ip_address[3],
-                               b""))
-    finally:
-        file.close()
-
-
-class LoginInfoReaderTest(LandscapeTest):
-    """Tests for login info file reader."""
-
-    def test_read_empty_file(self):
-        """Test ensures the reader is resilient to empty files."""
-        filename = self.makeFile("")
-
-        file = open(filename, "rb")
-        try:
-            reader = LoginInfoReader(file)
-            self.assertEqual(reader.read_next(), None)
-        finally:
-            file.close()
-
-    def test_read_login_info(self):
-        """Test ensures the reader can read login info."""
-        filename = self.makeFile("")
-        append_login_data(filename, login_type=1, pid=100, tty_device="/dev/",
-                          id="1", username="jkakar", hostname="localhost",
-                          termination_status=0, exit_status=0, session_id=1,
-                          entry_time_seconds=105, entry_time_milliseconds=10,
-                          remote_ip_address=[192, 168, 42, 102])
-        append_login_data(filename, login_type=1, pid=101, tty_device="/dev/",
-                          id="1", username="root", hostname="localhost",
-                          termination_status=0, exit_status=0, session_id=2,
-                          entry_time_seconds=235, entry_time_milliseconds=17,
-                          remote_ip_address=[192, 168, 42, 102])
-
-        file = open(filename, "rb")
-        try:
-            reader = LoginInfoReader(file)
-
-            info = reader.read_next()
-            self.assertEqual(info.login_type, 1)
-            self.assertEqual(info.pid, 100)
-            self.assertEqual(info.tty_device, "/dev/")
-            self.assertEqual(info.id, "1")
-            self.assertEqual(info.username, "jkakar")
-            self.assertEqual(info.hostname, "localhost")
-            self.assertEqual(info.termination_status, 0)
-            self.assertEqual(info.exit_status, 0)
-            self.assertEqual(info.session_id, 1)
-            self.assertEqual(info.entry_time, datetime.utcfromtimestamp(105))
-            # FIXME Test IP address handling. -jk
-
-            info = reader.read_next()
-            self.assertEqual(info.login_type, 1)
-            self.assertEqual(info.pid, 101)
-            self.assertEqual(info.tty_device, "/dev/")
-            self.assertEqual(info.id, "1")
-            self.assertEqual(info.username, "root")
-            self.assertEqual(info.hostname, "localhost")
-            self.assertEqual(info.termination_status, 0)
-            self.assertEqual(info.exit_status, 0)
-            self.assertEqual(info.session_id, 2)
-            self.assertEqual(info.entry_time, datetime.utcfromtimestamp(235))
-            # FIXME Test IP address handling. -jk
-
-            info = reader.read_next()
-            self.assertEqual(info, None)
-        finally:
-            file.close()
-
-    def test_login_info_iterator(self):
-        """Test ensures iteration behaves correctly."""
-        filename = self.makeFile("")
-        append_login_data(filename)
-        append_login_data(filename)
-
-        file = open(filename, "rb")
-        try:
-            reader = LoginInfoReader(file)
-            count = 0
-
-            for info in reader.login_info():
-                count += 1
-
-            self.assertEqual(count, 2)
-        finally:
-            file.close()
 
 
 class ComputerUptimeTest(LandscapeTest):
@@ -309,16 +203,3 @@ class ComputerUptimeTest(LandscapeTest):
         plugin.run()
         self.mstore.set_accepted_types(["computer-uptime"])
         self.assertMessages(list(self.mstore.get_pending_messages()), [])
-
-
-class BootTimesTest(LandscapeTest):
-
-    def test_fallback_to_uptime(self):
-        """
-        When no data is available in C{/var/log/wtmp}
-        L{BootTimes.get_last_boot_time} falls back to C{/proc/uptime}.
-        """
-        wtmp_filename = self.makeFile("")
-        append_login_data(wtmp_filename, tty_device="~", username="shutdown",
-                          entry_time_seconds=535)
-        self.assertTrue(BootTimes(filename=wtmp_filename).get_last_boot_time())
